chore: Update AppFixtures to fix date and duration calculations
Update the AppFixtures class to fix the calculation of the start date and duration for each sortie. The start date is now set to a random date between two weeks in the future, and the duration is a random number between 1 and 5 hours. Additionally, the end date is now correctly calculated by adding the duration to the start date. Finally, the etat (state) of the sortie is set to "Clôturée" if the current date is after the start date, otherwise it is set to "Créée" or "Ouverte".